+++ src/main/java/roomescape/controller/api/ReservationTimeApiController.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.controller.api;
 
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Positive;
